feat: 이벤트모달관련기능 hook화
